test: EnumForceSInt8
