+++ src/c/tests/enums/_index.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "EnumForceSInt8.h"
 #include "EnumForceUInt8.h"
 #include "EnumForceUInt16.h"
 #include "EnumForceUInt32.h"
