+++ src/cs/tests/C2CS.Tests/C/Data/Values/aarch64-apple-ios/Enums/EnumForceSInt8.json
@@ -1,0 +1,34 @@
+{
+  "name": "EnumForceSInt8",
+  "type_integer": "unsigned int",
+  "values": [
+    {
+      "name": "ENUM_FORCE_SINT8_DAY_UNKNOWN",
+      "value": 0
+    },
+    {
+      "name": "ENUM_FORCE_SINT8_DAY_MONDAY",
+      "value": 1
+    },
+    {
+      "name": "ENUM_FORCE_SINT8_DAY_TUESDAY",
+      "value": 2
+    },
+    {
+      "name": "ENUM_FORCE_SINT8_DAY_WEDNESDAY",
+      "value": 3
+    },
+    {
+      "name": "ENUM_FORCE_SINT8_DAY_THURSDAY",
+      "value": 4
+    },
+    {
+      "name": "ENUM_FORCE_SINT8_DAY_FRIDAY",
+      "value": 5
+    },
+    {
+      "name": "_ENUM_FORCE_SINT8",
+      "value": 127
+    }
+  ]
+}
